feat: isolate and track forester worker concurrency

+++ prover/server/prover/common/types.go
@@ -14,7 +14,7 @@ const (
 // JSON input structures (these are not in circuit_utils.go)
 type InclusionProofInputsJSON struct {
 	Root         string   `json:"root"`
-	PathIndex    uint32   `json:"pathIndex"`
+	PathIndex    uint64   `json:"pathIndex"`
 	PathElements []string `json:"pathElements"`
 	Leaf         string   `json:"leaf"`
 }
