+++ prover/server/prover/v2/inclusion_proving_system.go
@@ -16,7 +16,7 @@ import (
 
 type InclusionInputs struct {
 	Root            big.Int
-	PathIndex       uint32
+	PathIndex       uint64
 	PathElements    []big.Int
 	Leaf            big.Int
 	PublicInputHash big.Int
